Move blinx helper functions into dedicated file
